+++ backends/memcheck/ceed-memcheck.h
@@ -17,6 +17,16 @@ typedef struct {
   bool setup_done;
 } CeedQFunction_Memcheck;
 
+typedef struct {
+  int mem_block_id;
+  void *data;
+  void *data_allocated;
+  void *data_borrowed;
+  void *data_owned;
+} CeedQFunctionContext_Memcheck;
+
 CEED_INTERN int CeedQFunctionCreate_Memcheck(CeedQFunction qf);
+
+CEED_INTERN int CeedQFunctionContextCreate_Memcheck(CeedQFunctionContext ctx);
 
 #endif // _ceed_memcheck_h